// frontend-study-buddy/src/api/types.ts
export type UUID = string;

export type TipoUsuario = "UNIVERSITARIO" | "LOCADOR" | "ESTUDANTE";

export type Genero =
  | "MASCULINO"
  | "FEMININO"
  | "OUTRO"
  | "PREFIRO_NAO_INFORMAR";

export type TipoMoradia = "APARTAMENTO" | "CASA" | "REPUBLICA" | "QUARTO";

export type TipoAnuncio = "IMOVEL_COMPLETO" | "VAGA_COMPARTILHADA";

export type StatusAnuncio = "ATIVO" | "PAUSADO" | "ENCERRADO";
// ...
export interface AnuncioResponse {
  id: UUID;
  titulo: string;
  descricao: string;
  valorMensal: number;
  tipoAnuncio: TipoAnuncio;
  status: StatusAnuncio;
  dataPublicacao: string;
  anuncianteId: UUID;
  moradiaId: UUID;
}